requirements.txt
discord.py
pytz
langchain>0.0.244
openai
python-dotenv
PyNaCl
google-search-results
pymongo
dnspython
hupper
aiohttp
